bin/hackney/requirements.txt
astroid==2.3.3
backcall==0.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
decorator==4.4.2
google-api-python-client==1.7.12
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.17.0
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pylint==2.4.4
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.11.2
